pyproject.toml
[tool.poetry]
name = "testtaskrayapp"
version = "0.1.0"
description = ""
authors = ["BityutskiyNA"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.7"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
httpx = "^0.25.2"
schedule = "^1.2.1"
redis = "^5.0.1"
drf-spectacular = "^0.26.5"
pytest-django = "^4.7.0"
pytest-factoryboy = "^2.6.0"
gunicorn = "^21.2.0"
pydantic = {extras = ["dotenv"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
black = "^23.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
